@joyclee I hit the `raw` button from https://github.com/joyclee/datastories/blob/master/README.md hit copy.
Then I went to the about page, went to "edit this page in prose" button at the bottom, then went back to gallery, pages links, hit new file, and pasted it in.

Also changed the header, and renamed the file to `manifesto.html.md`.  Feel free to shoot me questions.  Be brave and make changes!